refactor: 파싱 관련 dto 패키지 이동 (#31)
